requirements.txt
comm_ddp==0.0.2
datasets==2.12.0
evaluate==0.4.0
imageio==2.31.1
joblib==1.2.0
matplotlib==3.7.1
nltk==3.8.1
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.5.3
peft==0.4.0
Pillow==10.0.0
pycocoevalcap==1.2
pydprint==0.0.4
Random_Word==1.0.11
Requests
rouge_score==0.1.2
setuptools==59.6.0
tensorflow==2.12.0
termcolor
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
wandb